Move directoryservice into accounts, add a little documentation to create and fix password fields to use password type
Ticket:#18113
